update review module

+++ src/main/java/com/vinhSeo/BookingCinema/service/ReviewService.java
@@ -16,7 +16,6 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -29,39 +28,30 @@ public class ReviewService {
     private final UserRepository userRepository;
     private final MovieRepository movieRepository;
 
-    @PreAuthorize("hasAnyAuthority('USER')")
-    public Review createReview(ReviewRequest request) {
-        log.info("Create review request by userId: {}", request.getUser());
+    public Review createReview(Integer userId, ReviewRequest request) {
+        log.info("Create review request by userId: {}", userId);
 
         Movie movie = movieRepository.findById(request.getMovie()).orElseThrow(() ->
                 new AppException(ErrorApp.MOVIE_NOT_FOUND));
 
-        User user = userRepository.findById(request.getUser()).orElseThrow(() ->
+        User user = userRepository.findById(userId).orElseThrow(() ->
                 new AppException(ErrorApp.USER_NOT_FOUND));
 
         if(reviewRepository.existsByUserAndMovie(user, movie)) {
             throw new AppException(ErrorApp.REVIEW_EXISTED);
         }
 
-        Review review = reviewMapper.toReview(request, userRepository, movieRepository);
-
-        log.info("id: {}", review.getId());
-        log.info("rating: {}", review.getRating());
-        log.info("comment: {}", review.getComment());
-        log.info("idUser: {}", review.getUser().getId());
-        log.info("idMovie: {}", review.getMovie().getId());
+        Review review = reviewMapper.toReview(request, movieRepository);
 
         return reviewRepository.save(review);
     }
 
-    @PreAuthorize("hasAnyAuthority('ADMIN', 'USER')")
     public Review getReviewById(Integer id) {
         log.info("Get review by id: {}", id);
 
         return reviewRepository.findById(id).orElseThrow(() -> new AppException(ErrorApp.REVIEW_NOT_FOUND));
     }
 
-    @PreAuthorize("hasAnyAuthority('ADMIN', 'USER')")
     public Page<Review> getAllByUser(Integer userId, Integer number, Integer size, String sortBy, String order) {
         log.info("Get all reviews by userId: {}", userId);
 
@@ -74,7 +64,6 @@ public class ReviewService {
         return reviewRepository.findAllByUser(user, pageable);
     }
 
-    @PreAuthorize("hasAnyAuthority('ADMIN', 'USER')")
     public Page<Review> getAllByMovie(Integer movieId, Integer number, Integer size, String sortBy, String order) {
         log.info("Get all reviews by movieId: {}", movieId);
 
@@ -87,29 +76,29 @@ public class ReviewService {
         return reviewRepository.findAllByMovie(movie, pageable);
     }
 
-    @PreAuthorize("hasAnyAuthority('USER')")
-    public Review updateReview(Integer id, ReviewRequest request) {
-        log.info("Update review request by userId: {}", request.getUser());
+    public Review updateReview(Integer id, Integer userId, ReviewRequest request) {
+        log.info("Update review request by id: {}", id);
 
         Review review = getReviewById(id);
-        ReviewRequest newRequest = ReviewRequest.builder()
-                .rating(request.getRating())
-                .comment(request.getComment())
-                .user(review.getUser().getId())
-                .movie(review.getMovie().getId())
-                .build();
 
-        Review newReview = reviewMapper.toReview(newRequest, userRepository, movieRepository);
-        newReview.setId(id);
+        if(!review.getUser().getId().equals(userId)) {
+            throw new AppException(ErrorApp.REVIEW_ACCESS_DENIED);
+        }
 
-        return reviewRepository.save(newReview);
+        review.setRating(request.getRating());
+        review.setComment(request.getComment());
+
+        return reviewRepository.save(review);
     }
 
-    @PreAuthorize("hasAnyAuthority('USER')")
-    public void deleteReview(Integer id) {
+    public void deleteReview(Integer id, Integer userId) {
         log.info("Delete review by id: {}", id);
 
-        getReviewById(id);
+        Review review = getReviewById(id);
+
+        if(!review.getUser().getId().equals(userId)) {
+            throw new AppException(ErrorApp.REVIEW_ACCESS_DENIED);
+        }
 
         reviewRepository.deleteById(id);
     }
